rename api and endshift, move a few more tests
